feat: refactor duration to java.time.Duration and align integer widths
Replace computed `durationMinutes: Int` on SleepLog with `duration: Duration`
for richer time arithmetic, add the same computed property to NewSleepLog,
and widen averageDurationMinutes to Long (int64) across model and OpenAPI spec.

+++ sleep/src/main/kotlin/com/noom/interview/fullstack/sleep/model/SleepStats.kt
@@ -15,7 +15,7 @@ import java.time.OffsetDateTime
 data class SleepStats(
     val dateFrom: OffsetDateTime,
     val dateTo: OffsetDateTime,
-    val averageDurationMinutes: Int?,
+    val averageDurationMinutes: Long?,
     val averageBedTime: LocalTime?,
     val averageWakeTime: LocalTime?,
     val moodFrequencies: Map<Mood, Int>
